Refactor ocr3 package in CLD
