+++ deployment/cre/ocr3/ocr3_1/changeset/operations/contracts/configure_vault_plugin.go
@@ -1,4 +1,4 @@
-package changeset
+package contracts
 
 import (
 	"encoding/hex"
@@ -13,12 +13,12 @@ import (
 	"github.com/smartcontractkit/chainlink-common/pkg/capabilities/actions/vault"
 	cldf "github.com/smartcontractkit/chainlink-deployments-framework/deployment"
 	"github.com/smartcontractkit/chainlink-deployments-framework/operations"
-
 	changesetstate "github.com/smartcontractkit/chainlink/deployment/common/changeset/state"
 	"github.com/smartcontractkit/chainlink/deployment/cre/common/strategies"
 	crecontracts "github.com/smartcontractkit/chainlink/deployment/cre/contracts"
 	"github.com/smartcontractkit/chainlink/deployment/cre/jobs/pkg"
-	"github.com/smartcontractkit/chainlink/deployment/cre/ocr3"
+	"github.com/smartcontractkit/chainlink/deployment/cre/ocr3/ocr3_1"
+	"github.com/smartcontractkit/chainlink/deployment/cre/ocr3/v2/changeset"
 	"github.com/smartcontractkit/chainlink/deployment/cre/ocr3/v2/changeset/operations/contracts"
 )
 
@@ -34,7 +34,7 @@ type ConfigureVaultPluginInput struct {
 	ContractQualifier     string `json:"contractQualifier" yaml:"contractQualifier"`
 
 	DON                   contracts.DonNodeSet         `json:"don" yaml:"don"`
-	OracleConfig          *ocr3.V3_1OracleConfig       `json:"oracleConfig" yaml:"oracleConfig"`
+	OracleConfig          *ocr3_1.V3_1OracleConfig     `json:"oracleConfig" yaml:"oracleConfig"`
 	DryRun                bool                         `json:"dryRun" yaml:"dryRun"`
 	InstanceID            InstanceIDComponents         `json:"instanceID" yaml:"instanceID"`
 	ReportingPluginConfig *vault.ReportingPluginConfig `json:"reportingPluginConfig,omitempty" yaml:"reportingPluginConfig,omitempty"`
@@ -60,23 +60,9 @@ func (l ConfigureVaultPlugin) VerifyPreconditions(_ cldf.Environment, input Conf
 	if input.OracleConfig == nil {
 		return errors.New("oracle config is required")
 	}
-	if !((input.OracleConfig.PrevConfigDigest == "") == (input.OracleConfig.PrevSeqNr == uint64(0)) && (input.OracleConfig.PrevSeqNr == uint64(0)) == (input.OracleConfig.PrevHistoryDigest == "")) { //nolint:staticcheck // ignore
-		return errors.New("PrevConfigDigest, PrevSeqNr, and PrevHistoryDigest must all be set or all be nil")
-	}
-	if (input.OracleConfig.PrevConfigDigest != "") && (input.OracleConfig.PrevSeqNr == uint64(0)) {
-		return errors.New("PrevSeqNr must be positive if PrevConfigDigest is set")
-	}
-	if input.OracleConfig.PrevConfigDigest != "" {
-		err := verify32BytesHexString(input.OracleConfig.PrevConfigDigest)
-		if err != nil {
-			return fmt.Errorf("invalid OracleConfig.PrevConfigDigest, should be hex encoded 32 bytes: %w", err)
-		}
-	}
-	if input.OracleConfig.PrevHistoryDigest != "" {
-		err := verify32BytesHexString(input.OracleConfig.PrevHistoryDigest)
-		if err != nil {
-			return fmt.Errorf("invalid OracleConfig.PrevHistoryDigest, should be hex encoded 32 bytes: %w", err)
-		}
+	_, _, err := ocr3_1.VerifyAndExtractOCR3_1Fields(input.OracleConfig.PrevConfigDigest, input.OracleConfig.PrevSeqNr, input.OracleConfig.PrevHistoryDigest)
+	if err != nil {
+		return errors.New("verifyAndExtractOCR3_1Fields failed verification: " + err.Error())
 	}
 
 	if input.InstanceID.DKGContractQualifier == "" {
@@ -85,7 +71,7 @@ func (l ConfigureVaultPlugin) VerifyPreconditions(_ cldf.Environment, input Conf
 	if input.InstanceID.ConfigDigest == "" {
 		return errors.New("instanceID.config_digest is required")
 	}
-	err := verify32BytesHexString(input.InstanceID.ConfigDigest)
+	_, err = ocr3_1.HexStringTo32ByteArray(input.InstanceID.ConfigDigest)
 	if err != nil {
 		return fmt.Errorf("invalid instanceID.configDigest, should be hex encoded 32 bytes: %w", err)
 	}
@@ -122,7 +108,7 @@ func (l ConfigureVaultPlugin) Apply(e cldf.Environment, input ConfigureVaultPlug
 		input.MCMSConfig,
 		mcmsContracts,
 		common.HexToAddress(contractAddrRef.Address),
-		ConfigureOCR3Description,
+		changeset.ConfigureOCR3Description,
 	)
 	if err != nil {
 		return cldf.ChangesetOutput{}, fmt.Errorf("failed to create strategy: %w", err)
@@ -147,10 +133,10 @@ func (l ConfigureVaultPlugin) Apply(e cldf.Environment, input ConfigureVaultPlug
 		return cldf.ChangesetOutput{}, fmt.Errorf("failed to marshal VaultPlugin reporting plugin config: %w", err)
 	}
 
-	report, err := operations.ExecuteOperation(e.OperationsBundle, contracts.ConfigureOCR3_1, contracts.ConfigureOCR3_1Deps{
+	report, err := operations.ExecuteOperation(e.OperationsBundle, ConfigureOCR3_1, ConfigureOCR3_1Deps{
 		Env:      &e,
 		Strategy: strategy,
-	}, contracts.ConfigureOCR3_1Input{
+	}, ConfigureOCR3_1Input{
 		ContractAddress:               &contractAddr,
 		ChainSelector:                 input.ContractChainSelector,
 		DON:                           input.DON,
@@ -167,15 +153,4 @@ func (l ConfigureVaultPlugin) Apply(e cldf.Environment, input ConfigureVaultPlug
 		MCMSTimelockProposals: report.Output.MCMSTimelockProposals,
 		Reports:               []operations.Report[any, any]{report.ToGenericReport()},
 	}, nil
-}
-
-func verify32BytesHexString(s string) error {
-	b, err := hex.DecodeString(s)
-	if err != nil {
-		return fmt.Errorf("failed to decode hex string: %w", err)
-	}
-	if len(b) != 32 {
-		return fmt.Errorf("hex string must be 32 bytes, got %d", len(b))
-	}
-	return nil
 }
